+++ src/main/meters/Correlometer.cpp
@@ -1,6 +1,6 @@
 /*
- * Copyright (C) 2024 Linux Studio Plugins Project <https://lsp-plug.in/>
- *           (C) 2024 Vladimir Sadovnikov <sadko4u@gmail.com>
+ * Copyright (C) 2025 Linux Studio Plugins Project <https://lsp-plug.in/>
+ *           (C) 2025 Vladimir Sadovnikov <sadko4u@gmail.com>
  *
  * This file is part of lsp-dsp-units
  * Created on: 8 мар. 2024 г.
@@ -86,9 +86,9 @@ namespace lsp
 
             vInA        = advance_ptr_bytes<float>(ptr, szof_buf);
             vInB        = advance_ptr_bytes<float>(ptr, szof_buf);
-            nCapacity   = capacity;
+            nCapacity   = uint32_t(capacity);
             nHead       = 0;
-            nMaxPeriod  = max_period;
+            nMaxPeriod  = uint32_t(max_period);
             nPeriod     = 0;
             nFlags      = 0;
 
@@ -106,7 +106,7 @@ namespace lsp
             if (period == nPeriod)
                 return;
 
-            nPeriod     = period;
+            nPeriod     = uint32_t(period);
             nFlags     |= CF_UPDATE;
         }
 
@@ -159,9 +159,9 @@ namespace lsp
                 const size_t can_do = nPeriod - nWindow;
                 size_t to_do = lsp_min(
                     count - offset,                             // Number of samples left in input buffers
-                    nCapacity - nMaxPeriod,                     // Number of free samples in the ring buffer
-                    nCapacity - nHead,                          // The number of samples before head goes out of ring buffer
-                    nCapacity - tail);                          // The number of samples before tail goes out of ring buffer
+                    size_t(nCapacity - nMaxPeriod),             // Number of free samples in the ring buffer
+                    size_t(nCapacity - nHead),                  // The number of samples before head goes out of ring buffer
+                    size_t(nCapacity - tail));                  // The number of samples before tail goes out of ring buffer
                 to_do = lsp_min(to_do, can_do);
 
                 // Fill buffers with data
